+++ test/browser.js
@@ -51,7 +51,7 @@ describe('Browser specific tests', () => {
   })
 
   describe('migrations tests', () => {
-    require('./migrations/migration-8-test')(() => createRepo(repoOptions), repoCleanup, repoOptions)
+    require('./migrations')(() => createRepo(repoOptions), repoCleanup, repoOptions)
   })
 
   describe('init tests', () => {